Feature/recent events (#17)

+++ src/bot/handlers/events/events.hpp
@@ -6,6 +6,6 @@ namespace NeverForgetBot::Events {
 
     void send_event(TgBot::Bot& bot, const Event& event);
 
-    void send_events(TgBot::Bot &bot, long long chat_id, const std::vector<Event> &events);
+    void send_events(TgBot::Bot &bot, long chat_id, const std::vector<Event> &events);
 
 } // namespace NeverForgetBot::Events
